Fix references to ServiceStartCommand and Options as they are ServerStart
