+++ core/Azure.Mcp.Core/src/Services/Azure/AzureServiceCollectionExtensions.cs
@@ -39,12 +39,12 @@ public static class AzureServiceCollectionExtensions
     public static IServiceCollection AddAzureService(this IServiceCollection services)
     {
         // !!! HACK !!!
-        // Program.cs for the CLI servers have their own DI containers vs ServiceStartCommand.
+        // Program.cs for the CLI servers have their own DI containers vs ServerStartCommand.
         // If the CLI is started to run a non-"server start" command, then we're assuming we should
         // use the identity resolved from the host environment for downstream auth (e.g., Azure CLI
         // or VS Code user). This will fulfill the DI container for those non-"server start" commands.
-        // Within ServiceStartCommand, when it has its own fully IHost with DI and IConfiguration,
-        // then it will need to call AddAzureTokenCredentialProvider for just the ServiceStartCommand
+        // Within ServerStartCommand, when it has its own fully IHost with DI and IConfiguration,
+        // then it will need to call AddAzureTokenCredentialProvider for just the ServerStartCommand
         // container to be populated with the correct authentication strategy, such as OBO for
         // running as a remote HTTP MCP service.
         services.AddSingleIdentityTokenCredentialProvider();
