+++ core/Azure.Mcp.Core/tests/Azure.Mcp.Core.Tests/Areas/Server/ServerStartCommandLiveTests.cs
@@ -1,0 +1,743 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Diagnostics;
+using Microsoft.Mcp.Tests;
+using Microsoft.Mcp.Tests.Client.Helpers;
+using ModelContextProtocol.Client;
+using ModelContextProtocol.Protocol;
+using Xunit;
+
+namespace Azure.Mcp.Core.Tests.Areas.Server;
+
+/// <summary>
+/// Live integration tests for Azure MCP Server that validate tool loading behavior across different modes.
+/// These tests start actual MCP server instances and verify the correct tools are loaded.
+/// </summary>
+[Trait("TestType", "Live")]
+public class ServerStartCommandLiveTests(ITestOutputHelper output) : IAsyncLifetime
+{
+    private Process? _httpServerProcess;
+
+    protected ITestOutputHelper Output { get; } = output;
+
+    public ValueTask InitializeAsync()
+    {
+        Assert.SkipWhen(!TestExtensions.IsLiveTestMode(), "Skipping test in non-live mode");
+        return ValueTask.CompletedTask;
+    }
+
+    public ValueTask DisposeAsync()
+    {
+        if (_httpServerProcess is { HasExited: false } process)
+        {
+            try
+            {
+                process.Kill(entireProcessTree: true);
+                process.Dispose();
+            }
+            catch (InvalidOperationException)
+            {
+                // Process already exited
+            }
+        }
+
+        _httpServerProcess = null;
+
+        return ValueTask.CompletedTask;
+    }
+
+    private async Task<McpClient?> CreateClientAsync(params string[] arguments)
+    {
+        string executablePath = McpTestUtilities.GetAzMcpExecutablePath();
+
+        LiveTestSettings.TryLoadTestSettings(out var settings);
+        Dictionary<string, string?> envVars = settings?.EnvironmentVariables.ToDictionary(k => k.Key, v => (string?)v.Value) ?? [];
+
+        var (client, serverUrl) = await McpTestUtilities.CreateMcpClientAsync(
+            executablePath,
+            [.. arguments],
+            envVars,
+            process => _httpServerProcess = process,
+            Output,
+            settings?.TestPackage,
+            settings?.SettingsDirectory);
+
+        return client;
+    }
+
+    #region Default Mode Tests
+
+    [Fact]
+    public async Task DefaultMode_LoadsNamespaceTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        // Should have tools from multiple areas
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Default mode is now namespace mode, so should have namespace-level tools (not 60+ individual tools)
+        Assert.True(toolNames.Count > 20, $"Expected more than 20 namespace tools, got {toolNames.Count}");
+
+        // Should include the documentation tool (displayed by its title)
+        Assert.Contains("documentation", toolNames, StringComparer.OrdinalIgnoreCase);
+
+        // Should include subscription and group utility commands
+        Assert.Contains(toolNames, name => name.Contains("subscription", StringComparison.OrdinalIgnoreCase) && name.Contains("list", StringComparison.OrdinalIgnoreCase));
+        Assert.Contains(toolNames, name => name.Contains("group", StringComparison.OrdinalIgnoreCase) && name.Contains("list", StringComparison.OrdinalIgnoreCase));
+
+        var uniqueNames = toolNames.Distinct().ToList();
+
+        Assert.Equal(toolNames.Count, uniqueNames.Count);
+
+        Output.WriteLine($"Verified {toolNames.Count} unique tool names in default (namespace) mode");
+
+        // Log for debugging
+        Output.WriteLine($"Default mode (namespace) loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    [Fact]
+    public async Task DefaultMode_CanCallSubscriptionList()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start");
+
+        // Act
+        var result = await client!.CallToolAsync("subscription_list", new Dictionary<string, object?> { },
+            cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+        Assert.NotNull(result.Content);
+        Assert.NotEmpty(result.Content);
+
+        // The result should contain subscription data (even if empty list)
+        var firstContent = result.Content.FirstOrDefault();
+        Assert.NotNull(firstContent);
+
+        // Log for debugging
+        Output.WriteLine($"Subscription list result: {firstContent}");
+    }
+
+    [Fact]
+    public async Task DefaultMode_CanCallGroupList()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start");
+
+        // Act
+        var result = await client!.CallToolAsync("group_list", new Dictionary<string, object?> { },
+            cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+        Assert.NotNull(result.Content);
+        Assert.NotEmpty(result.Content);
+
+        // The result should contain resource group data (even if empty list)
+        var firstContent = result.Content.FirstOrDefault();
+        Assert.NotNull(firstContent);
+
+        // Log for debugging
+        Output.WriteLine($"Group list result: {firstContent}");
+    }
+
+    #endregion
+
+    #region All Mode Tests
+
+    [Fact]
+    public async Task AllMode_LoadsAllIndividualTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "all");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        // Should have tools from multiple areas
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should include Azure service tools and extension tools (all individual tools)
+        Assert.True(toolNames.Count > 60, $"Expected more than 60 individual tools, got {toolNames.Count}");
+
+        // Should include the microsoft_docs_search tool
+        Assert.Contains("microsoft_docs_search", toolNames, StringComparer.OrdinalIgnoreCase);
+
+        // All tool names should be unique.
+        var uniqueNames = toolNames.Distinct().ToList();
+        Assert.Equal(toolNames.Count, uniqueNames.Count);
+
+        Output.WriteLine($"Verified {toolNames.Count} unique tool names in all mode");
+
+        // Log for debugging
+        Output.WriteLine($"All mode loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    #endregion
+
+    #region Single Tool Proxy Mode Tests
+
+    [Fact]
+    public async Task SingleProxyMode_LoadsSingleAzureTool()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "single");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.Single(listResult);
+
+        var tool = listResult.First();
+        Assert.Equal("azure", tool.Name);
+        Assert.Contains("azure", tool.Description, StringComparison.OrdinalIgnoreCase);
+
+        Output.WriteLine($"Single proxy mode loaded 1 tool: {tool.Name}");
+        Output.WriteLine($"Description: {tool.Description}");
+    }
+
+    [Fact]
+    public async Task SingleProxyMode_WithNamespaceFilter_StillLoadsSingleAzureTool()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "single", "--namespace", "storage");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        // Single proxy mode should still expose single azure tool regardless of namespace filter
+        Assert.Single(listResult);
+        Assert.Equal("azure", listResult.First().Name);
+
+        Output.WriteLine("Single proxy mode with namespace filter still loaded 1 tool");
+    }
+
+    [Fact]
+    public async Task SingleProxyMode_WithReadOnlyFlag_LoadsSingleAzureTool()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "single", "--read-only");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.Single(listResult);
+        Assert.Equal("azure", listResult.First().Name);
+
+        // Verify read-only behavior for single proxy mode
+        var tool = listResult.First();
+        var readOnlyHint = tool.ProtocolTool?.Annotations?.ReadOnlyHint;
+
+        // The azure tool should have read-only hint when in read-only mode
+        Output.WriteLine($"Single proxy read-only tool ReadOnlyHint: {readOnlyHint}");
+        Output.WriteLine($"Tool annotations: {tool.ProtocolTool?.Annotations != null}");
+
+        // For single proxy mode, the ReadOnlyHint may be null because it's a proxy tool
+        // but we should still have annotations present
+        Assert.NotNull(tool.ProtocolTool?.Annotations);
+
+        Output.WriteLine("Single proxy read-only mode loaded 1 tool");
+    }
+
+    #endregion
+
+    #region Namespace Proxy Mode Tests
+
+    [Fact]
+    public async Task NamespaceProxyMode_LoadsNamespaceTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "namespace");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // In namespace mode without specific namespaces, should default to extension tools
+        Assert.True(toolNames.Count > 20, "Should have more than 20 tools in namespace mode");
+
+        // Should include the documentation tool (displayed by its title)
+        Assert.Contains("documentation", toolNames, StringComparer.OrdinalIgnoreCase);
+
+        Output.WriteLine($"Namespace proxy mode loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    [Fact]
+    public async Task NamespaceProxyMode_WithSpecificNamespaces_LoadsNamespaceSpecificTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync(
+            "server", "start",
+            "--mode", "namespace",
+            "--namespace", "storage",
+            "--namespace", "keyvault",
+            "--namespace", "documentation");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should include namespace-specific tools
+        var hasRelevantTools = toolNames.Any(name =>
+            name.Contains("documentation", StringComparison.OrdinalIgnoreCase) ||
+            name.Contains("storage", StringComparison.OrdinalIgnoreCase) ||
+            name.Contains("keyvault", StringComparison.OrdinalIgnoreCase));
+
+        Assert.True(hasRelevantTools, "Should have tools related to specified namespaces");
+
+        // Should contain exactly 6 tools: 3 specified namespaces + 3 utility tools (group_list, group_resource_list, subscription_list)
+        Assert.Equal(6, toolNames.Count);
+
+        // Verify tools are from storage, keyvault namespaces, or utility tools
+        Assert.All(toolNames, toolName =>
+        {
+            var isStorageOrKeyVault = toolName.Contains("documentation", StringComparison.OrdinalIgnoreCase) ||
+                toolName.Contains("storage", StringComparison.OrdinalIgnoreCase) ||
+                toolName.Contains("keyvault", StringComparison.OrdinalIgnoreCase);
+            var isUtilityTool = toolName.Contains("group_list", StringComparison.OrdinalIgnoreCase) ||
+                toolName.Contains("group_resource_list", StringComparison.OrdinalIgnoreCase) ||
+                toolName.Contains("subscription_list", StringComparison.OrdinalIgnoreCase);
+            Assert.True(isStorageOrKeyVault || isUtilityTool, $"Tool '{toolName}' should be related to documentation, keyvault, storage namespaces, or be a utility tool");
+        });
+
+        Output.WriteLine($"Namespace proxy mode with [documentation, keyvault, storage] loaded {toolNames.Count} tools");
+    }
+
+    [Fact]
+    public async Task NamespaceProxyMode_StorageToolLearnMode_ReturnsStorageCommands()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "namespace");
+
+        // Act - Call storage tool in learn mode
+        var learnParameters = new Dictionary<string, object?>
+        {
+            ["learn"] = true
+        };
+
+        var result = await client!.CallToolAsync("storage", learnParameters, cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+        Assert.NotNull(result.Content);
+        Assert.NotEmpty(result.Content);
+
+        // Get the text content
+        var textContent = result.Content.OfType<TextContentBlock>().FirstOrDefault();
+        Assert.NotNull(textContent);
+        Assert.NotEmpty(textContent.Text);
+
+        var responseText = textContent.Text;
+
+        // Verify the response contains information about storage commands
+        Assert.Contains("available command", responseText, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("storage", responseText, StringComparison.OrdinalIgnoreCase);
+
+        // Verify it contains specific storage commands we expect
+        Assert.Contains("storage_account_get", responseText, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("storage_blob_container_get", responseText, StringComparison.OrdinalIgnoreCase);
+
+        Output.WriteLine("Storage tool learn mode response:");
+        Output.WriteLine(responseText);
+        Output.WriteLine($"✓ Learn mode returned {responseText.Length} characters of storage command information");
+    }
+
+    #endregion
+
+    #region Default Mode with Filters Tests
+
+    [Fact]
+    public async Task DefaultMode_WithNamespaceFilter_LoadsFilteredTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--namespace", "storage", "--namespace", "keyvault");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+        Assert.Equal(5, listResult.Count()); // 2 specified namespaces + 3 utility tools
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should only include tools from specified namespaces
+        var hasStorageOrKeyVault = toolNames.Any(name =>
+            name.Contains("storage", StringComparison.OrdinalIgnoreCase) ||
+            name.Contains("keyvault", StringComparison.OrdinalIgnoreCase));
+
+        Assert.True(hasStorageOrKeyVault, "Should have tools related to storage or keyvault namespaces");
+
+        Output.WriteLine($"Default mode with namespaces [storage, keyvault] loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    [Fact]
+    public async Task AllMode_WithNamespaceFilter_LoadsFilteredIndividualTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "all", "--namespace", "storage", "--namespace", "keyvault");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should only include individual tools from specified namespaces
+        var hasStorageOrKeyVault = toolNames.Any(name =>
+            name.Contains("storage", StringComparison.OrdinalIgnoreCase) ||
+            name.Contains("keyvault", StringComparison.OrdinalIgnoreCase));
+
+        Assert.True(hasStorageOrKeyVault, "Should have tools related to storage or keyvault namespaces");
+
+        // In all mode with namespace filter, should have more individual tools than namespace mode
+        Assert.True(toolNames.Count > 2, $"Expected more than 2 individual tools, got {toolNames.Count}");
+
+        Output.WriteLine($"All mode with namespaces [storage, keyvault] loaded {toolNames.Count} tools");
+    }
+
+    [Fact]
+    public async Task AllMode_WithReadOnlyFlag_LoadsOnlyReadOnlyTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "all", "--read-only");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        // All tools should be read-only or the count should be reduced
+        var toolCount = listResult.Count();
+        Assert.True(toolCount > 0, "Should have at least some read-only tools");
+
+        // Verify tools have read-only annotations
+        var toolsWithReadOnlyHint = 0;
+        var toolsWithAnnotations = 0;
+
+        foreach (var tool in listResult)
+        {
+            var hasAnnotations = tool.ProtocolTool?.Annotations != null;
+            var readOnlyHint = tool.ProtocolTool?.Annotations?.ReadOnlyHint;
+
+            if (hasAnnotations)
+            {
+                toolsWithAnnotations++;
+            }
+
+            if (readOnlyHint.HasValue && readOnlyHint.Value)
+            {
+                toolsWithReadOnlyHint++;
+            }
+
+            Output.WriteLine($"Tool: {tool.Name} - HasAnnotations: {hasAnnotations}, ReadOnlyHint: {readOnlyHint}");
+        }
+
+        Output.WriteLine($"Tools with annotations: {toolsWithAnnotations}/{toolCount}");
+        Output.WriteLine($"Tools with ReadOnlyHint=true: {toolsWithReadOnlyHint}/{toolCount}");
+
+        // In read-only mode, ALL tools must have annotations and ReadOnlyHint=true
+        Assert.Equal(toolCount, toolsWithAnnotations);
+        Assert.Equal(toolCount, toolsWithReadOnlyHint);
+
+        // Additional verification message
+        Output.WriteLine("✓ All tools have annotations and ReadOnlyHint=true as expected in read-only mode");
+
+        Output.WriteLine($"Default read-only mode loaded {toolCount} tools");
+    }
+
+    #endregion
+
+    #region Negative Tests - Invalid Modes and Namespaces
+
+    [Fact]
+    public async Task InvalidMode_FailsToStartServer()
+    {
+        // Act & Assert
+        // The 2.0 SDK surfaces an early transport failure (the server process exits before the
+        // connection completes) as ClientTransportClosedException. The wrapped InnerException is
+        // platform-dependent (an IOException on Windows, null on Linux), so only the outer type
+        // is asserted.
+        await Assert.ThrowsAsync<ClientTransportClosedException>(async () =>
+        {
+            await using var client = await CreateClientAsync("server", "start", "--mode", "invalid-mode");
+        });
+    }
+
+    [Fact]
+    public async Task InvalidNamespace_LoadsGracefully()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--namespace", "invalid-namespace", "--namespace", "another-invalid");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        // Should not crash, but may have fewer tools
+        Assert.NotNull(listResult);
+
+        // Invalid namespaces should result in only utility tools (group_list, group_resource_list, subscription_list)
+        Assert.Equal(3, listResult.Count());
+        var toolNames = listResult.Select(t => t.Name).ToList();
+        Assert.Contains(toolNames, name => name.Contains("group", StringComparison.OrdinalIgnoreCase));
+        Assert.Contains(toolNames, name => name.Contains("subscription", StringComparison.OrdinalIgnoreCase));
+
+        Output.WriteLine($"Invalid namespaces loaded {listResult.Count()} tools");
+    }
+
+    #endregion
+
+    #region Consolidated Proxy Mode Tests
+
+    [Fact]
+    public async Task ConsolidatedProxyMode_LoadsConsolidatedTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "consolidated");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // In consolidated mode, should have consolidated tools grouping related operations
+        Assert.True(toolNames.Count > 20, $"Expected more than 20 consolidated tools, got {toolNames.Count}");
+
+        // Should include some known consolidated tools
+        Assert.Contains(toolNames, name => name.Contains("azure_subscriptions_and_resource_groups", StringComparison.OrdinalIgnoreCase));
+        Assert.Contains(toolNames, name => name.Contains("azure_databases_details", StringComparison.OrdinalIgnoreCase));
+        Assert.Contains(toolNames, name => name.Contains("azure_storage_details", StringComparison.OrdinalIgnoreCase));
+
+        Output.WriteLine($"Consolidated proxy mode loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    [Fact]
+    public async Task ConsolidatedProxyMode_ToolLearnMode_ReturnsConsolidatedCommands()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "consolidated");
+
+        // Act - Call a consolidated tool in learn mode
+        var learnParameters = new Dictionary<string, object?>
+        {
+            ["learn"] = true
+        };
+
+        var result = await client!.CallToolAsync("get_azure_databases_details", learnParameters, cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+        Assert.NotNull(result.Content);
+        Assert.NotEmpty(result.Content);
+
+        // Get the text content
+        var textContent = result.Content.OfType<TextContentBlock>().FirstOrDefault();
+        Assert.NotNull(textContent);
+        Assert.NotEmpty(textContent.Text);
+
+        var responseText = textContent.Text;
+
+        // Verify the response contains information about database commands
+        Assert.Contains("available command", responseText, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("database", responseText, StringComparison.OrdinalIgnoreCase);
+
+        // Verify it contains multiple database-related commands
+        Assert.Contains("mysql", responseText, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("postgres", responseText, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("sql", responseText, StringComparison.OrdinalIgnoreCase);
+
+        Output.WriteLine("Consolidated database tool learn mode response:");
+        Output.WriteLine(responseText);
+        Output.WriteLine($"✓ Learn mode returned {responseText.Length} characters of consolidated command information");
+    }
+
+    [Fact]
+    public async Task ConsolidatedProxyMode_WithNamespaceFilter_LoadsFilteredConsolidatedTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "consolidated", "--namespace", "storage");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should only include consolidated tools related to specified namespaces
+        var hasRelevantTools = toolNames.Any(name =>
+            name.Contains("storage", StringComparison.OrdinalIgnoreCase));
+
+        Assert.True(hasRelevantTools, "Should have consolidated tools related to storage namespaces");
+
+        // In consolidated mode with namespace filter, should have fewer tools than without filter
+        Assert.True(toolNames.Count < 10, $"Expected fewer than 10 tools with namespace filter, got {toolNames.Count}");
+
+        Output.WriteLine($"Consolidated proxy mode with [storage] namespaces loaded {toolNames.Count} tools");
+        foreach (var name in toolNames)
+        {
+            Output.WriteLine($"  - {name}");
+        }
+    }
+
+    [Fact]
+    public async Task ConsolidatedProxyMode_WithReadOnlyFlag_LoadsOnlyReadOnlyConsolidatedTools()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "consolidated", "--read-only");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolCount = listResult.Count();
+        Assert.True(toolCount > 0, "Should have at least some read-only consolidated tools");
+
+        // Verify all tools have read-only annotations
+        var toolsWithReadOnlyHint = 0;
+        var toolsWithAnnotations = 0;
+
+        foreach (var tool in listResult)
+        {
+            var hasAnnotations = tool.ProtocolTool?.Annotations != null;
+            var readOnlyHint = tool.ProtocolTool?.Annotations?.ReadOnlyHint;
+
+            if (hasAnnotations)
+            {
+                toolsWithAnnotations++;
+            }
+
+            if (readOnlyHint.HasValue && readOnlyHint.Value)
+            {
+                toolsWithReadOnlyHint++;
+            }
+
+            // Verify tool names don't contain destructive operations
+            Assert.DoesNotContain("create_", tool.Name, StringComparison.OrdinalIgnoreCase);
+            Assert.DoesNotContain("edit_", tool.Name, StringComparison.OrdinalIgnoreCase);
+            Assert.DoesNotContain("delete_", tool.Name, StringComparison.OrdinalIgnoreCase);
+            Assert.DoesNotContain("update_", tool.Name, StringComparison.OrdinalIgnoreCase);
+        }
+    }
+
+    [Fact]
+    public async Task ConsolidatedProxyMode_CanCallConsolidatedTool()
+    {
+        // Arrange
+        await using var client = await CreateClientAsync("server", "start", "--mode", "consolidated");
+
+        // Act - Call the consolidated subscriptions and resource groups tool
+        var result = await client!.CallToolAsync("get_azure_subscriptions_and_resource_groups",
+            new Dictionary<string, object?> { },
+            cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+        Assert.NotNull(result.Content);
+        Assert.NotEmpty(result.Content);
+
+        // The result should contain subscription and resource group data
+        var firstContent = result.Content.FirstOrDefault();
+        Assert.NotNull(firstContent);
+
+        // Log for debugging
+        Output.WriteLine($"Consolidated tool result: {firstContent}");
+    }
+
+    #endregion
+
+    #region Tool Mode Tests
+
+    [Fact]
+    public async Task ToolMode_AutomaticallyChangesToAllMode()
+    {
+        // Arrange - Test that --tool switch automatically changes mode to "all"
+        await using var client = await CreateClientAsync("server", "start", "--tool", "group_list", "--tool", "subscription_list");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should only include the specified tools
+        Assert.Equal(2, toolNames.Count);
+        Assert.Contains("group_list", toolNames);
+        Assert.Contains("subscription_list", toolNames);
+    }
+
+    [Fact]
+    public async Task ToolMode_OverridesExplicitNamespaceMode()
+    {
+        // Arrange - Test that --tool switch overrides --mode namespace
+        await using var client = await CreateClientAsync("server", "start", "--mode", "namespace", "--tool", "group_list");
+
+        // Act
+        var listResult = await client!.ListToolsAsync(cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotEmpty(listResult);
+
+        var toolNames = listResult.Select(t => t.Name).ToList();
+
+        // Should only include the specified tool, mode should be automatically changed to "all"
+        Assert.Single(toolNames);
+        Assert.Contains("group_list", toolNames);
+    }
+
+    #endregion
+}
