+++ core/Microsoft.Mcp.Core/src/Areas/Server/Commands/ServiceCollectionExtensions.cs
@@ -37,28 +37,28 @@ public static partial class ServiceCollectionExtensions
     /// Adds the Azure MCP server services to the specified <see cref="IServiceCollection"/>.
     /// </summary>
     /// <param name="services">The service collection to add services to.</param>
-    /// <param name="serviceStartOptions">The options for configuring the server.</param>
+    /// <param name="serverStartOptions">The options for configuring the server.</param>
     /// <returns>The service collection with MCP server services added.</returns>
-    public static IServiceCollection AddAzureMcpServer(this IServiceCollection services, ServerStartOptions serviceStartOptions)
+    public static IServiceCollection AddAzureMcpServer(this IServiceCollection services, ServerStartOptions serverStartOptions)
     {
         // Register HTTP client services
         services.AddHttpClientServices();
 
         // Register options for service start
-        services.AddSingleton(serviceStartOptions);
-        services.AddSingleton(Options.Create(serviceStartOptions));
+        services.AddSingleton(serverStartOptions);
+        services.AddSingleton(Options.Create(serverStartOptions));
 
         // Register default tool loader options from service start options
         var defaultToolLoaderOptions = new ToolLoaderOptions
         {
-            Namespace = serviceStartOptions.Namespace,
-            ReadOnly = serviceStartOptions.ReadOnly ?? false,
-            DangerouslyDisableElicitation = serviceStartOptions.DangerouslyDisableElicitation,
-            Tool = serviceStartOptions.Tool,
-            IsHttpMode = serviceStartOptions.IsHttpMode
+            Namespace = serverStartOptions.Namespace,
+            ReadOnly = serverStartOptions.ReadOnly ?? false,
+            DangerouslyDisableElicitation = serverStartOptions.DangerouslyDisableElicitation,
+            Tool = serverStartOptions.Tool,
+            IsHttpMode = serverStartOptions.IsHttpMode
         };
 
-        if (serviceStartOptions.Mode == ModeTypes.NamespaceProxy)
+        if (serverStartOptions.Mode == ModeTypes.NamespaceProxy)
         {
             if (defaultToolLoaderOptions.Namespace == null || defaultToolLoaderOptions.Namespace.Length == 0)
             {
@@ -88,7 +88,7 @@ public static partial class ServiceCollectionExtensions
         services.AddSingleton<IMcpRuntime, McpRuntime>();
 
         // Register MCP discovery strategies based on proxy mode
-        if (serviceStartOptions.Mode == ModeTypes.SingleToolProxy)
+        if (serverStartOptions.Mode == ModeTypes.SingleToolProxy)
         {
             services.AddSingleton<IMcpDiscoveryStrategy>(sp =>
             {
@@ -102,11 +102,11 @@ public static partial class ServiceCollectionExtensions
                 return new CompositeDiscoveryStrategy(discoveryStrategies, logger);
             });
         }
-        else if (serviceStartOptions.Mode == ModeTypes.NamespaceProxy)
+        else if (serverStartOptions.Mode == ModeTypes.NamespaceProxy)
         {
             services.AddSingleton<IMcpDiscoveryStrategy, RegistryDiscoveryStrategy>();
         }
-        else if (serviceStartOptions.Mode == ModeTypes.ConsolidatedProxy)
+        else if (serverStartOptions.Mode == ModeTypes.ConsolidatedProxy)
         {
             services.AddSingleton<IMcpDiscoveryStrategy>(sp =>
             {
@@ -122,11 +122,11 @@ public static partial class ServiceCollectionExtensions
         }
 
         // Configure tool loading based on mode
-        if (serviceStartOptions.Mode == ModeTypes.SingleToolProxy)
+        if (serverStartOptions.Mode == ModeTypes.SingleToolProxy)
         {
             services.AddSingleton<IToolLoader, SingleProxyToolLoader>();
         }
-        else if (serviceStartOptions.Mode == ModeTypes.NamespaceProxy)
+        else if (serverStartOptions.Mode == ModeTypes.NamespaceProxy)
         {
             services.AddSingleton<IToolLoader>(sp =>
             {
@@ -168,7 +168,7 @@ public static partial class ServiceCollectionExtensions
                 return new CompositeToolLoader(toolLoaders, loggerFactory.CreateLogger<CompositeToolLoader>());
             });
         }
-        else if (serviceStartOptions.Mode == ModeTypes.ConsolidatedProxy)
+        else if (serverStartOptions.Mode == ModeTypes.ConsolidatedProxy)
         {
             services.AddSingleton<IToolLoader>(sp =>
             {
@@ -198,7 +198,7 @@ public static partial class ServiceCollectionExtensions
                 return new CompositeToolLoader(toolLoaders, loggerFactory.CreateLogger<CompositeToolLoader>());
             });
         }
-        else if (serviceStartOptions.Mode == ModeTypes.All)
+        else if (serverStartOptions.Mode == ModeTypes.All)
         {
             services.AddSingleton<IMcpDiscoveryStrategy, RegistryDiscoveryStrategy>();
             services.AddSingleton<IToolLoader>(sp =>
@@ -241,7 +241,7 @@ public static partial class ServiceCollectionExtensions
 
         var mcpServerBuilder = services.AddMcpServer();
 
-        if (serviceStartOptions.Transport == TransportTypes.Http)
+        if (serverStartOptions.Transport == TransportTypes.Http)
         {
             mcpServerBuilder.WithHttpTransport();
         }
@@ -263,7 +263,7 @@ public static partial class ServiceCollectionExtensions
         services.AddSingleton(GetConfiguration());
 
         services.AddOptions<McpServerConfiguration>()
-            .Configure<IConfiguration, IOptions<ServerStartOptions>>((options, rootConfiguration, serviceStartOptions) =>
+            .Configure<IConfiguration, IOptions<ServerStartOptions>>((options, rootConfiguration, serverStartOptions) =>
             {
                 // Use a scoped IConfiguration for loading server settings.
                 var scopedConfiguration = GetConfiguration(assembly);
@@ -304,7 +304,7 @@ public static partial class ServiceCollectionExtensions
                 // Disable telemetry when support logging is enabled to prevent sensitive data from being sent
                 // to telemetry endpoints. Support logging captures debug-level information that may contain
                 // sensitive data, so we disable all telemetry as a safety measure.
-                if (!string.IsNullOrWhiteSpace(serviceStartOptions.Value.DangerouslyWriteSupportLogsToDir))
+                if (!string.IsNullOrWhiteSpace(serverStartOptions.Value.DangerouslyWriteSupportLogsToDir))
                 {
                     options.IsTelemetryEnabled = false;
                     return;
