+++ servers/Azure.Mcp.Server/src/Program.cs
@@ -330,7 +330,7 @@ internal class Program
         // !!! WARNING !!!
         // stdio-transport-specific implementations of IAzureService and ICacheService.
         // The http-transport-specific implementations and configurations must be registered
-        // within ServiceStartCommand.ExecuteAsync().
+        // within ServerStartCommand.ExecuteAsync().
         services.AddHttpClientServices(configureDefaults: true);
         services.AddAzureService();
         services.AddSingleUserCliCacheService(disabled: true);
